requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
tensorflow>=2.20.0
keras>=3.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dateutil>=2.8.2
pydantic>=2.4.0
pydantic-settings>=2.0.0
